refactor: move Result.ts and add comments
